# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dashboard Rokokgs API",
        "description": "RESTful API untuk Sistem Management Penjualan Rokok Gas - Dashboard Web dan Mobile App",
        "contact": {
            "email": "admin@rokokgs.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.rokokgs.com",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/api/v1/areas": {
            "get": {
                "tags": [
                    "Areas"
                ],
                "summary": "Get areas list",
                "description": "Sales melihat area yang di-assign ke mereka, Admin/Manager melihat semua area",
                "operationId": "70a1cc5dde7084c9732718c9eb12b4a7",
                "responses": {
                    "200": {
                        "description": "Areas retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Areas retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Jakarta Selatan"
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "JKT-S"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Wilayah Jakarta Selatan"
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/areas/{id}": {
            "get": {
                "tags": [
                    "Areas"
                ],
                "summary": "Get area by ID",
                "description": "Mendapatkan detail area berdasarkan ID",
                "operationId": "71dacbf3d544513effddaae4692a1a65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Area ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Area retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Area retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jakarta Selatan"
                                                },
                                                "code": {
                                                    "type": "string",
                                                    "example": "JKT-S"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Wilayah Jakarta Selatan"
                                                },
                                                "is_active": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Area not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Area not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token Authentication",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}